refactor(mcp): extract fastest-model tool modules
